+++ game/level.go
@@ -3,8 +3,8 @@ package game
 import (
 	"fmt"
 
-	"github.com/mokiat/lacking/async"
 	"github.com/mokiat/lacking/game/asset"
+	"github.com/mokiat/lacking/util/async"
 )
 
 type SceneDefinition struct {
